+++ packages/smarthr-ui/src/components/Dropdown/DropdownContentInner.tsx
@@ -5,6 +5,7 @@ import {
   type FC,
   type PropsWithChildren,
   createContext,
+  useContext,
   useEffect,
   useMemo,
   useRef,
@@ -12,11 +13,20 @@ import {
 } from 'react'
 import { tv } from 'tailwind-variants'
 
+import { useLatest } from '../../hooks/useLatest'
 import { useTheme } from '../../hooks/useTheme'
+import { tabbable } from '../../libs/tabbable'
 
+import { DropdownContext } from './Dropdown'
 import { DropdownCloser } from './DropdownCloser'
-import { type ContentBoxStyle, type Rect, getContentBoxStyle } from './dropdownHelper'
-import { useKeyboardNavigation } from './useKeyboardNavigation'
+import {
+  type ContentBoxStyle,
+  type Rect,
+  getContentBoxStyle,
+  getFirstTabbable,
+} from './dropdownHelper'
+
+const KEY_ESCAPE = /^Esc(ape)?$/
 
 const classNameGenerator = tv({
   base: 'smarthr-ui-Dropdown-content shr-absolute shr-z-overlap-base shr-overflow-y-auto shr-break-words shr-rounded-m shr-bg-white shr-shadow-layer-3 forced-colors:shr-outline forced-colors:shr-outline-1',
@@ -120,7 +130,97 @@ export const DropdownContentInner: FC<Props> = ({
     }
   }, [isActive])
 
-  useKeyboardNavigation(wrapperRef, focusTargetRef)
+  const { triggerElementRef, rootTriggerRef, handleDelegateClickCloser } =
+    useContext(DropdownContext)
+
+  const latest = useLatest({
+    triggerElementRef,
+    rootTriggerRef,
+    handleDelegateClickCloser,
+  })
+
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key === 'Tab') {
+        if (
+          !wrapperRef.current ||
+          !latest.triggerElementRef.current ||
+          !latest.rootTriggerRef?.current
+        ) {
+          return
+        }
+
+        const tabbablesInContent = tabbable(wrapperRef.current)
+
+        if (tabbablesInContent.length === 0) {
+          return
+        }
+
+        const trigger = tabbable(latest.triggerElementRef.current).at(-1)
+        const firstTabbable = tabbablesInContent[0]
+
+        if (e.target === trigger) {
+          if (e.shiftKey) {
+            // move focus previous of the Trigger
+            return
+          }
+
+          // focus a first tabbable element in the dropdown content
+          e.preventDefault()
+          firstTabbable.focus()
+
+          return
+        } else if (e.shiftKey) {
+          if (e.target === firstTabbable || e.target === focusTargetRef.current) {
+            // focus the Trigger
+            e.preventDefault()
+            trigger!.focus()
+            latest.handleDelegateClickCloser()
+          }
+        } else if (e.target === tabbablesInContent.at(-1)) {
+          // move focus next of the Trigger
+          const rootTrigger = tabbable(latest.rootTriggerRef.current).at(-1)
+
+          if (rootTrigger) {
+            rootTrigger.focus()
+            latest.handleDelegateClickCloser()
+          }
+        }
+      } else if (KEY_ESCAPE.test(e.key)) {
+        if (e.target && e.target === focusTargetRef.current) {
+          latest.handleDelegateClickCloser()
+
+          return
+        }
+
+        const trigger = getFirstTabbable(latest.triggerElementRef)
+
+        if (trigger && e.target === trigger) {
+          // close the dropdown when the Trigger is focused and Esc key is pressed
+          latest.handleDelegateClickCloser()
+
+          return
+        }
+
+        if (wrapperRef.current) {
+          for (const inner of tabbable(wrapperRef.current)) {
+            if (inner === e.target) {
+              // close the dropdown when an element that is included in dropdown content is focused and Esc key is pressed
+              latest.handleDelegateClickCloser()
+
+              break
+            }
+          }
+        }
+      }
+    }
+
+    window.addEventListener('keydown', handleKeyDown)
+
+    return () => {
+      window.removeEventListener('keydown', handleKeyDown)
+    }
+  }, [latest])
 
   return (
     <div {...rest} ref={wrapperRef} className={actualClassName} style={style}>
